Login & logout, recent activity, Rooms CRUD, Comment, Search functionalities are added.

+++ base/views.py
@@ -1,7 +1,65 @@
 from django.db.models import Q
+from django.http import HttpResponse
 from django.shortcuts import render, redirect
-from .models import Room, Topic
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from .models import Room, Topic, Message
 from .forms import RoomForm
+
+def loginPage(request):
+    page = 'login'
+    
+    if request.user.is_authenticated:
+        return redirect('home')
+    
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+        
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist!')
+            
+        user = authenticate(request, username=username, password=password)
+        
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or Password does not exist!') 
+            
+    context = {
+        'page': page
+    }
+    return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm()
+    
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occured during registration!')
+    
+    context = {
+        'form': form
+    }
+    
+    return render(request, 'base/login_register.html', context)
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -13,19 +71,38 @@ def home(request):
     
     topics = Topic.objects.all()
     room_count = rooms.count()
+    room_messages = Message.objects.all()
     
     context = {
         'rooms': rooms,
         'topics': topics,
-        'room_count': room_count
+        'room_count': room_count,
+        'room_messages': room_messages
     }
     return render(request, 'base/home.html', context)
 
 def room(request, pk):
     room = Room.objects.get(pk=pk)
-    context = {'room': room}
+    room_messages = room.message_set.all().order_by('-created_at')
+    participants = room.participants.all()
+    
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user = request.user,
+            room = room,
+            body = request.POST.get('body')
+        )
+        room.participants.add(request.user)
+        return redirect('room', pk=room.id)
+    
+    context = {
+        'room': room,
+        'room_messages': room_messages,
+        'participants': participants
+    }
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     
@@ -38,9 +115,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(pk=pk)
     form = RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -51,11 +132,29 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(pk=pk)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
     
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     
-    return render(request, 'base/delete.html', {'room': room})
+    return render(request, 'base/delete.html', {'object': room})
+
+@login_required(login_url='login')
+def deleteMessage(request, pk):
+    message = Message.objects.get(pk=pk)
+    
+    if request.user != message.user:
+        return HttpResponse('You are not allowed here!')
+    
+    if request.method == 'POST':
+        message.delete()
+        return redirect('room', pk=message.room.id)
+    
+    return render(request, 'base/delete.html', {'object': message})
+
